refactor(bigquery): reorg dataset integration tests
